iteration F1: putting the cart in the sidebar In order to put the cart in the sidebar, we broke the cart view into a partial. We also turned the line items into a partial inside cart, so we could abstract away a bit of the logic.
After turning cart into a partial we could add it to the sidebar, but had to
update it's style when inside the sidebar. To do that we used the id cart and
nested styles in scss.

We also had to update the StoreController, so it would be able to get and
show our cart.

And also had to add an if to our application.html.erb, so other pages won't
break if there isn't a cart. It is only rendered when one exists. Our tests
were able to get this bug.