iteration f2: creating an ajax based cart
To do so, we started by adding remote: true to the button 'add to cart'
call,this way an ajax request is sent.

Instead of redirecting to index, we must update the HTML fragment,
to do so, we need to respond with a js format. So we added format.js in
our line_items_controller #create.

Now we need the js file it's redirected to, so we created a create.js.erb.
Thats where our #create action takes us. There we find the HTML we want to
update(the div with the id cart) and update it(rendering the cart object).