iteration f3: highlighting changes
To highlight that our cart was modified we added a background color animation.
In order for the animation to run, it needs to know if the current item is the
same as the line item.

So the create method in line_items_controller needs to pass that information
to our view, in this case a partial. That is why { @current_item = @line_item }
is added.

After that our partial needs to check whether or not it should add the class
that runs the animation. If the current and line item are the same, it displays
the <tr> with class highlight, otherwise it displays the one without a class.

We also added a test to our line items controller. xhr stands for the XMLHttpRequest