iteration f4: hiding an empty cart
we made an helper to only render the cart if it has an item. this helper
is called from the application view.

we also made an ajax call to hide the notice after it was displayed. before
it was only changed when the page refreshed, since our cart is ajax, the
notice would appear onde the cart was emptied, but would remain there even
after we added a new item (since the page wasnt refreshed)