iteration f5: updates with ActionCable
To enable real time updates – when a price is edited it immediattelly is
updated in the store - we added ActionCable. It integrates the WebSocket
protocol to our app.

We create a channel, broadcast data, and receive data.

We generate a channel products, update it to stream from products, update
the update action in products controller.

Since I have a visit counter, I also had to add it to the update action,
before when I edited the price I would get an error because @count was nil.
Updating it in the action to equate sessions[:counter] solved that. And the
counter is only incremented when the page is refreshed.

We ordered the product by title to make sure that when the page is updated
the order remais the same, so the update is seamless.

Since this update happens in the browser we use javascript to update our div
main with class store, with the new data we received.