decrement line_item in cart
I added the decrement action  in the line item controller, and the router.
I used a button_to to invoke the action.

It took a while to get done because I always seemed to forget something, for
instance setting the cart and line item.

It does work, however it renders the cart where the store should be rendered,
a small fix. I will work on that.