capturing an order
In order to create an order we generated the scaffold and added it's
reference to line_item.

This was a pretty big iteration. We had to complete our Order model and
controllers, also had to change tests, and views. It was quite a lot. I
think some important things were:

- Use of enums in order model
- Form helpers interact with the controller and with the models to
implement an integrated solution for form handling. Because the text
field is constructed in the context of form_with, it's automatically
associated with the data in the order object. This association means
that submitting the form will set the right names and values in the data
available to the controller, but it will also pre-populate the form fields
with any values already existing on the model
- Model objects perform two roles: they map data into and out of the
database, but they're also regular objects that hold business data. They
affect the database only when you tell them to, typically by calling save

Things that need to be fixed:
- I have two tests that aren't passing due to playtime exercises
– I added price in the playtime, but it isn't showing in the db, something
is off
– My form scss seems to be a bit wonky when it fails validations