migrated the line_items prices again
so, I'm pretty sure this worked the first time, because I had checked the
prices to make sure they were there. But now they were all null.

I made a new migration with the same code, except for creating the column,
since it is already there. And all the prices were successfully copied.

I think I need to keep an eye on this one, in case it breaks again. I might
be missing something.