G2: build atom feeds
We use atom feeds to build a 'proto' admin interface in a way.

We build an action in the products_controller that looks for a atom template.
In that template we can see all the orders of a product, and it's details.
If we have an action, we also need a route.

Important:
- feeds: the responses contain a bit of metadata that identifies when
the content was last modified and a hashed value called an ETag. If a
subsequent request provides this data back, this gives the server the
opportunity to respond with an empty response body and an indication that
the data hasn't been modified.
- Through: Products and orders have no direct relationship to each other,
though there is an indirect one via line items. The has_many method we used
to tell Rails that a product has many line items takes an optional argument
named through: that tells Rails to traverse the indirect relationship.
- Route: This action will respond to HTTP GET requests and will operate
on a member of the collection (in other words, on an individual product)
as opposed to the entire collection (which in this case would mean all
products)