Iteration H1: WIP
We are going to work with react in order to render the response to the
chosen pay type in the checkout. That way we only need to change a fraction
of our view, and don't need to call our server.

In order to do that we install react through the webpacker gem. Webpacker
has webpack all set up for us. We used the javascript_pack_tag helper method
in order to connect to our pack file.

The code in our pay_type pack looks a bit different because ReactDOM.render
is legacy, now we use createRoot(element).render(ui).

I also changed the import to react-dom/client because of a warning, but
it's apparently a bug that is gonna be fixed soon, so import from ReactDom
should work without any warnings soon.

By Rails convetion, we created a folder PayTypeSelector with an index file
in it, that is our React Component. Wepack is configured to look for folders
and files like this. We want to use the component in this module in our pack,
so we export it.

In our component we have chosen the name value for select in the same way
a Rails helper would, this way our controller is able to find the values,
even if they are coming from a React component, and not a Rails view.