iteration H1:  Adding Fields Dynamically to a Form
In this iteration we learned how to install react with the webpacker gem,
why use it and how to work with it in Rails.

Why
By using javascript we can render only parts of our app that need to be
changed, instead of the whole app, which makes the user experience faster.
And using React will make it even faster and easier than using vanilla js.
The core concept in React is component. A component is a view and a state.
We track state to change the view.

How
We created a new pack, that is the root of our implementation (pay_type).
We created a PayTypeSelector folder with our main component(index) and
the parts of it to be changed in our PayTypeCustomComponent inside our
index. We export the "parts"from their files, and import them in the index.
We add a constructor to our main component, where we can bind what needs
binding (anything we want to make sure has the correct value when .this
is called), and we can initialize state. We use the javascript_pack_tag
to connect our root to our rails view. The fields' names values are
named according to Rails convetion, so we can access those.