iterationI: added mailer
I added the mailer, but have a problem with the line_item partial I'm
using because of the delete item button. I'm trying to figure out how
to use a different partial in the mailer. But because I have some -
bigger than I would like - intervals between moments working on this,
I forget how some parts are connected..