Connecting to a Slow Payment Processor with Active Job
We have implemented the integration with the payment processor
(Pago, a fake API) as a method (charge) inside Order. Then the
orders_controller uses Active Job to execute charge in a
background job.

Why use a background job? So that the user doesn't wait for a response
that might take long, and is not relevant to the user experience. And
to avoid timeouts.

We used OpenStruct to create our fake API Pago. It makes an object
that responds to the methods given to its constructor.

We moved our charge logic from the orders_controller to the Order
model, because it became too complex. Our method:
– adapts pay_type_params to the ones Pago requires
– makes the call to Pago
– checks if the payment succeeded
– sends the confirmation email

To make our background job we created an ActiveJob class, that is
called from the orders_controller using the method perform_later,
and then calls charge.