fixed? the bugs, passed the system test
after i updated the yarn package in the master (when trying to run
rails s there it asked), I merge it to the iterationI branch, and
that seems to have fixed things.

My yarn has been acting up, it's not the first time I had problems
with it (recently). So I should keep an eye on that.