a notification email is sent when the ship_date is updated
for now I'm checking in the controller if the ship_date
was updated, if so I sent the new email I created to
send a notification with the date.

Tales sugested doing this in the model, using Active Model
Dirty: https://api.rubyonrails.org/classes/ActiveModel/Dirty.html
I want to look into it in the near future.